[CB-4480] Using 64 bit ints to store file size
Thanks to Klaus Schmidt for the original patch.

+++ CordovaLib/Classes/CDVFileTransfer.h
@@ -60,7 +60,7 @@ extern NSString* const kOptionsKeyCookie;
 
 @interface CDVFileTransferDelegate : NSObject {}
 
-- (void)updateBytesExpected:(NSInteger)newBytesExpected;
+- (void)updateBytesExpected:(long long)newBytesExpected;
 - (void)cancelTransfer:(NSURLConnection*)connection;
 
 @property (strong) NSMutableData* responseData; // atomic
@@ -73,8 +73,8 @@ extern NSString* const kOptionsKeyCookie;
 @property (nonatomic, copy) NSString* target;
 @property (nonatomic, copy) NSString* mimeType;
 @property (assign) int responseCode; // atomic
-@property (nonatomic, assign) NSInteger bytesTransfered;
-@property (nonatomic, assign) NSInteger bytesExpected;
+@property (nonatomic, assign) long long bytesTransfered;
+@property (nonatomic, assign) long long bytesExpected;
 @property (nonatomic, assign) BOOL trustAllHosts;
 @property (strong) NSFileHandle* targetFileHandle;
 @property (nonatomic, strong) CDVFileTransferEntityLengthRequest* entityLengthRequest;
